finish re-structure service folder
