+++ apps/mfe-shell-angular/src/app/features/expenses/report/report.component.ts
@@ -18,7 +18,7 @@ import {
   Subject,
   switchMap,
 } from 'rxjs';
-import { ExpenseService } from '../../../services/ExpenseService/expense.service';
+import { ExpenseService } from '@micro-expense-tracker/expenses/data-access';
 import {
   ExpenseList,
   FilterParams,
