--- apps/mfe-shell-angular/src/app/services/UserService/user-service.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { UserServiceService } from './user-service.service';
-
-describe('UserServiceService', () => {
-  let service: UserServiceService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(UserServiceService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
