--- apps/mfe-shell-angular/src/app/services/UserService/user-service.service.ts
@@ -1,19 +1,0 @@
-import { inject, Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
-import { from, Observable } from 'rxjs';
-import { CurrencyAPIObject } from '../../modal/settings-modal/settings-modal.component';
-
-@Injectable({
-  providedIn: 'root',
-})
-export class UserServiceService {
-  api = 'https://open.er-api.com/v6/latest/';
-
-  private http = inject(HttpClient);
-
-  constructor() {}
-
-  getCurrencyExchange(baseCurrency: string): Observable<CurrencyAPIObject> {
-    return from(this.http.get<CurrencyAPIObject>(`${this.api}${baseCurrency}`));
-  }
-}
