+++ libs/expenses/data-access/README.md
@@ -1,0 +1,3 @@
+# expenses-data-access
+
+This library was generated with [Nx](https://nx.dev).
