+++ libs/expenses/data-access/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/expenses-data-access/expense.service';
