+++ libs/shared/data-access/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/shared-data-access/ThemeService/theme.service';
+export * from './lib/shared-data-access/LocalStorage/local-storage.service';
