+++ libs/shared/types/src/index.ts
@@ -2,6 +2,7 @@ export * from './lib/interface/types';
 export * from './lib/interface/user.interface';
 export * from './lib/interface/shared.interface';
 export * from './lib/interface/chart.interface';
-export * from './lib/interface/expense.interface';
+export * from '../../../expenses/data-access/src/lib/expenses-data-access/expense.interface';
 export * from './lib/interface/modal.interface';
 export * from './lib/interface/settings.interface';
+export * from './lib/interface/nav.interface';
