+++ libs/shared/types/src/lib/interface/modal.interface.ts
@@ -1,4 +1,4 @@
-import { EditExpense } from './expense.interface';
+import { EditExpense } from '../../../../../expenses/data-access/src/lib/expenses-data-access/expense.interface';
 
 export interface DialogData {
   title: string;
