+++ libs/shared/types/src/lib/interface/nav.interface.ts
@@ -1,0 +1,7 @@
+export enum NavItem {
+  DASHBOARD = 'dashboard',
+  REPORT = 'report',
+  EXPENSE = 'expense',
+  USER = 'user',
+  MESSAGE = 'message',
+}
