+++ libs/shared/types/src/lib/interface/shared.interface.ts
@@ -3,3 +3,9 @@ import { DarkModeToggleProps } from '@micro-expense-tracker/shared/types';
 export interface ReactComponentType {
   default: React.ComponentType<DarkModeToggleProps>;
 }
+
+export interface FilterParams {
+  searchTerm?: string;
+  month?: number;
+  year?: number;
+}
