+++ libs/shared/ui/src/lib/ui/components/sidebar/sidebar.component.ts
@@ -10,12 +10,11 @@ import { MatSidenavModule } from '@angular/material/sidenav';
 import { AuthService } from '@micro-expense-tracker/auth/data-access';
 import { CommonModule } from '@angular/common';
 import { MatIcon } from '@angular/material/icon';
-import { DialogActionEnum } from '../../../interface/modal.interface';
+import { DialogActionEnum, NavItem } from '@micro-expense-tracker/shared/types';
 import { MatDialog } from '@angular/material/dialog';
 import { faArrowRightFromBracket } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
-import { navItems } from '../../../common/common-list';
-import { NavItem } from '../../../interface/expense.interface';
+import { navItems } from '@micro-expense-tracker/shared/constants';
 import { RouterModule, Router, NavigationEnd } from '@angular/router';
 import { filter } from 'rxjs/operators';
 import { ReactWrapperComponent } from '../react-wrapper/react-wrapper.component';
